config refactor avoiding circular dep issues
