--- modules/network/base.ts
@@ -1,230 +1,0 @@
-import { BigNumber, ethers } from 'ethers';
-import { DeploymentEnv, NetworkConfig, NetworkData } from './network-config-types';
-import { tokenService } from '../token/token.service';
-import { BoostedPoolAprService } from '../pool/lib/apr-data-sources/boosted-pool-apr.service';
-import { SwapFeeAprService } from '../pool/lib/apr-data-sources/swap-fee-apr.service';
-import { GaugeAprService } from '../pool/lib/apr-data-sources/ve-bal-gauge-apr.service';
-import { GaugeStakingService } from '../pool/lib/staking/gauge-staking.service';
-import { BptPriceHandlerService } from '../token/lib/token-price-handlers/bpt-price-handler.service';
-import { LinearWrappedTokenPriceHandlerService } from '../token/lib/token-price-handlers/linear-wrapped-token-price-handler.service';
-import { SwapsPriceHandlerService } from '../token/lib/token-price-handlers/swaps-price-handler.service';
-import { UserSyncGaugeBalanceService } from '../user/lib/user-sync-gauge-balance.service';
-import { every } from '../../worker/intervals';
-import { GithubContentService } from '../content/github-content.service';
-import { gaugeSubgraphService } from '../subgraphs/gauge-subgraph/gauge-subgraph.service';
-import { CoingeckoPriceHandlerService } from '../token/lib/token-price-handlers/coingecko-price-handler.service';
-import { coingeckoService } from '../coingecko/coingecko.service';
-import { env } from '../../app/env';
-import { IbTokensAprService } from '../pool/lib/apr-data-sources/ib-tokens-apr.service';
-
-const baseNetworkData: NetworkData = {
-    chain: {
-        slug: 'base',
-        id: 8453,
-        nativeAssetAddress: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
-        wrappedNativeAssetAddress: '0x4200000000000000000000000000000000000006',
-        prismaId: 'BASE',
-        gqlId: 'BASE',
-    },
-    subgraphs: {
-        startDate: '2023-07-10',
-        balancer: 'https://api.studio.thegraph.com/query/24660/balancer-base-v2/version/latest',
-        beetsBar: '',
-        blocks: 'https://api.studio.thegraph.com/query/48427/bleu-base-blocks/version/latest',
-        gauge: 'https://api.studio.thegraph.com/query/24660/balancer-gauges-base/version/latest',
-        veBalLocks: 'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-gauges',
-        userBalances: '',
-    },
-    eth: {
-        address: '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee',
-        addressFormatted: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
-        symbol: 'ETH',
-        name: 'Ether',
-    },
-    weth: {
-        address: '0x4200000000000000000000000000000000000006',
-        addressFormatted: '0x4200000000000000000000000000000000000006',
-    },
-    coingecko: {
-        nativeAssetId: 'ethereum',
-        platformId: 'base',
-        excludedTokenAddresses: [],
-    },
-    tokenPrices: {
-        maxHourlyPriceHistoryNumDays: 100,
-    },
-    rpcUrl: 'https://base.gateway.tenderly.co/7mM7DbBouY1JjnQd9MMDsd',
-    rpcMaxBlockRange: 500,
-    sanity: {
-        projectId: '',
-        dataset: '',
-    },
-    protocolToken: 'bal',
-    bal: {
-        address: '0x4158734d47fc9692176b5085e0f52ee0da5d47f1',
-    },
-    veBal: {
-        address: '0xc128a9954e6c874ea3d62ce62b468ba073093f25',
-        delegationProxy: '0xd87f44df0159dc78029ab9ca7d7e57e7249f5acd',
-    },
-    balancer: {
-        vault: '0xba12222222228d8ba445958a75a0704d566bf2c8',
-        composableStablePoolFactories: ['0x8df317a729fcaa260306d7de28888932cb579b88'],
-        weightedPoolV2Factories: ['0x4c32a8a8fda4e24139b51b456b42290f51d6a1c4'],
-        swapProtocolFeePercentage: 0.5,
-        yieldProtocolFeePercentage: 0.5,
-    },
-    ibAprConfig: {
-        defaultHandlers: {
-            cbETH: {
-                tokenAddress: '0x2ae3f1ec7f1f5012cfeab0185bfc7aa3cf0dec22',
-                sourceUrl: 'https://api.exchange.coinbase.com/wrapped-assets/CBETH/',
-                path: 'apy',
-                scale: 1,
-                isIbYield: true,
-            },
-        },
-    },
-    multicall: '0xca11bde05977b3631167028862be2a173976ca11',
-    multicall3: '0xca11bde05977b3631167028862be2a173976ca11',
-    avgBlockSpeed: 2,
-    sor: {
-        main: {
-            url: 'https://uu6cfghhd5lqa7py3nojxkivd40zuugb.lambda-url.ca-central-1.on.aws/',
-            maxPools: 8,
-            forceRefresh: false,
-            gasPrice: BigNumber.from(10),
-            swapGas: BigNumber.from('1000000'),
-        },
-        canary: {
-            url: 'https://ksa66wlkjbvteijxmflqjehsay0jmekw.lambda-url.eu-central-1.on.aws/',
-            maxPools: 8,
-            forceRefresh: false,
-            gasPrice: BigNumber.from(10),
-            swapGas: BigNumber.from('1000000'),
-        },
-    },
-    monitoring: {
-        main: {
-            alarmTopicArn: 'arn:aws:sns:ca-central-1:118697801881:api_alarms',
-        },
-        canary: {
-            alarmTopicArn: 'arn:aws:sns:eu-central-1:118697801881:api_alarms',
-        },
-    },
-};
-
-export const baseNetworkConfig: NetworkConfig = {
-    data: baseNetworkData,
-    contentService: new GithubContentService(),
-    provider: new ethers.providers.JsonRpcProvider({ url: baseNetworkData.rpcUrl, timeout: 60000 }),
-    poolAprServices: [
-        new IbTokensAprService(baseNetworkData.ibAprConfig),
-        new BoostedPoolAprService(),
-        new SwapFeeAprService(baseNetworkData.balancer.swapProtocolFeePercentage),
-        new GaugeAprService(tokenService, [baseNetworkData.bal!.address]),
-    ],
-    poolStakingServices: [new GaugeStakingService(gaugeSubgraphService, baseNetworkData.bal!.address)],
-    tokenPriceHandlers: [
-        new CoingeckoPriceHandlerService(coingeckoService),
-        new BptPriceHandlerService(),
-        new LinearWrappedTokenPriceHandlerService(),
-        new SwapsPriceHandlerService(),
-    ],
-    userStakedBalanceServices: [new UserSyncGaugeBalanceService()],
-    /*
-    For sub-minute jobs we set the alarmEvaluationPeriod and alarmDatapointsToAlarm to 1 instead of the default 3. 
-    This is needed because the minimum alarm period is 1 minute and we want the alarm to trigger already after 1 minute instead of 3.
-
-    For every 1 days jobs we set the alarmEvaluationPeriod and alarmDatapointsToAlarm to 1 instead of the default 3. 
-    This is needed because the maximum alarm evaluation period is 1 day (period * evaluationPeriod).
-    */
-    workerJobs: [
-        {
-            name: 'update-token-prices',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(10, 'minutes') : every(2, 'minutes'),
-        },
-        {
-            name: 'update-liquidity-for-inactive-pools',
-            interval: every(1, 'days'),
-            alarmEvaluationPeriod: 1,
-            alarmDatapointsToAlarm: 1,
-        },
-        {
-            name: 'update-liquidity-for-active-pools',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(6, 'minutes') : every(2, 'minutes'),
-        },
-        {
-            name: 'update-pool-apr',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(6, 'minutes') : every(2, 'minutes'),
-        },
-        {
-            name: 'load-on-chain-data-for-pools-with-active-updates',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(4, 'minutes') : every(1, 'minutes'),
-        },
-        {
-            name: 'sync-new-pools-from-subgraph',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(6, 'minutes') : every(2, 'minutes'),
-        },
-        {
-            name: 'sync-tokens-from-pool-tokens',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(10, 'minutes') : every(5, 'minutes'),
-        },
-        {
-            name: 'update-liquidity-24h-ago-for-all-pools',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(10, 'minutes') : every(5, 'minutes'),
-        },
-        {
-            name: 'cache-average-block-time',
-            interval: every(1, 'hours'),
-        },
-        {
-            name: 'sync-staking-for-pools',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(10, 'minutes') : every(5, 'minutes'),
-        },
-        {
-            name: 'sync-latest-snapshots-for-all-pools',
-            interval: every(1, 'hours'),
-        },
-        {
-            name: 'update-lifetime-values-for-all-pools',
-            interval: every(30, 'minutes'),
-        },
-        {
-            name: 'sync-changed-pools',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(2, 'minutes') : every(20, 'seconds'),
-            alarmEvaluationPeriod: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? 3 : 1,
-            alarmDatapointsToAlarm: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? 3 : 1,
-        },
-        {
-            name: 'user-sync-wallet-balances-for-all-pools',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(30, 'minutes') : every(10, 'minutes'),
-        },
-        {
-            name: 'user-sync-staked-balances',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(30, 'minutes') : every(10, 'minutes'),
-        },
-        {
-            name: 'sync-coingecko-coinids',
-            interval: every(2, 'hours'),
-        },
-        {
-            name: 'purge-old-tokenprices',
-            interval: every(1, 'days'),
-            alarmEvaluationPeriod: 1,
-            alarmDatapointsToAlarm: 1,
-        },
-        {
-            name: 'update-fee-volume-yield-all-pools',
-            interval: every(1, 'hours'),
-        },
-        {
-            name: 'sync-vebal-balances',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(9, 'minutes') : every(3, 'minutes'),
-        },
-        {
-            name: 'sync-vebal-totalSupply',
-            interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(10, 'minutes') : every(5, 'minutes'),
-        },
-    ],
-};
